refactor(#211): move private methods into subsystem classes, shrink game.h
Move private method declarations from Game onto their subsystem owners:
- GameRenderer: SetColors, palette fade, SetupRenderDevice, TimerCallback
- GameInput: UpdateInputState, scancode handlers, gamepad, TickRumble/Nav
- GameUiPipeline: Clay frame pipeline methods
- GameSession: LoadMap, UnloadGame, Check* helpers, GiveDefaultItems

Create game_ui_pipeline.cpp and game_session.cpp.
