+++ clients/silencer/src/client/ui/screens/screen_context.cpp
@@ -50,5 +50,5 @@ void ScreenContext::ShowMessage(const char * msg, std::function<void()> onClose)
 void ScreenContext::ResetPresentation(int paletteIdx) {
 	renderer.palette.SetPalette(paletteIdx);
 	game.GetScreenBuffer().Clear(0);
-	game.SetColors(renderer.palette.GetColors());
+	game.gameRenderer.SetColors(renderer.palette.GetColors());
 }
