+++ clients/silencer/src/game/game.cpp
@@ -21,26 +21,6 @@ void Game::Present() {
 gameRenderer.Present();
 }
 
-bool Game::SetupRenderDevice() {
-return gameRenderer.Setup(&gameRenderer.WindowRef());
-}
-
-void Game::SetColors(SDL_Color * colors) {
-gameRenderer.SetColors(colors);
-}
-
-void Game::UpdateInputState(Input & input) {
-gameInput.UpdateInputState(input);
-}
-
-bool Game::LoadMap(const char * name) {
-return gameSession.LoadMap(name);
-}
-
-void Game::UnloadGame() {
-gameSession.UnloadGame();
-}
-
 void Game::JoinGame(LobbyGame & lobbygame, char * password) {
 gameSession.JoinGame(lobbygame, password);
 }
@@ -51,62 +31,6 @@ gameSession.SpectateGame(lobbygame, password);
 
 void Game::LeaveJoinedGame() {
 gameSession.LeaveJoinedGame();
-}
-
-void Game::ShowDeployMessage() {
-gameSession.ShowDeployMessage();
-}
-
-void Game::GiveDefaultItems(Player & player) {
-gameSession.GiveDefaultItems(player);
-}
-
-void Game::RestartPaletteFade() {
-gameRenderer.RestartPaletteFade();
-}
-
-float Game::LegacyUiAnimationStepSeconds() const {
-return gameRenderer.LegacyUiAnimationStepSeconds();
-}
-
-Uint8 Game::PaletteFadePhaseFromClock() const {
-return gameRenderer.PaletteFadePhaseFromClock();
-}
-
-bool Game::PaletteFadeFinished() const {
-return gameRenderer.PaletteFadeFinished();
-}
-
-void Game::ApplyPaletteFade(bool fadeOut) {
-gameRenderer.ApplyPaletteFade(fadeOut);
-}
-
-void Game::PrepareClientUiFrame(Surface& surface) {
-gameUiPipeline.PrepareFrame(surface);
-}
-
-void Game::BeginPreparedClientUiFrame() {
-gameUiPipeline.BeginFrame();
-}
-
-Clay_RenderCommandArray Game::EndClientUiFrame() {
-return gameUiPipeline.EndFrame();
-}
-
-void Game::RenderClientUiFrame(Surface& surface, float frametime) {
-gameUiPipeline.RenderFrame(surface, frametime);
-}
-
-void Game::ResetUiFrameDeltas() {
-gameUiPipeline.ResetDeltas();
-}
-
-void Game::BuildVisibleClientUi(Surface& surface, float frametime) {
-gameUiPipeline.BuildVisible(surface, frametime);
-}
-
-void Game::DrawInGameWorldInsets(Surface& surface, float frametime) {
-gameUiPipeline.DrawInGameWorldInsets(surface, frametime);
 }
 
 void Game::PushScreen(std::unique_ptr<Screen> s) {
@@ -127,40 +51,4 @@ return gameUiPipeline.Top();
 
 bool Game::HasUiInputTarget() {
 return gameUiPipeline.HasInputTarget();
-}
-
-void Game::OpenFirstGamepad() {
-gameInput.OpenFirstGamepad();
-}
-
-void Game::PollGamepadState() {
-gameInput.PollGamepadState();
-}
-
-void Game::OnScancodeDown(int scancode) {
-gameInput.OnScancodeDown(scancode);
-}
-
-void Game::OnScancodeUp(int scancode) {
-gameInput.OnScancodeUp(scancode);
-}
-
-void Game::QueueUiKeyboardInputForScancode(int scancode) {
-gameUiPipeline.QueueKeyboardInputForScancode(
-scancode,
-gameInput.GetKeystate(),
-gameInput.GetKeyMap(),
-gameInput.GetGamepadState());
-}
-
-void Game::TickGamepadMenuNav() {
-gameInput.TickGamepadMenuNav();
-}
-
-void Game::TickRumble() {
-gameInput.TickRumble(world.GetPeerPlayer(world.localpeerid));
-}
-
-const char * Game::GetActionKeyDisplayName(Action a) {
-return gameInput.GetActionKeyDisplayName(a);
 }
