+++ clients/silencer/src/game/render/game_render.cpp
@@ -1,6 +1,6 @@
 #include "game.h"
 
-Uint32 Game::TimerCallback(void * userdata, SDL_TimerID timerID, Uint32 interval){
+Uint32 GameRenderer::TimerCallback(void * userdata, SDL_TimerID timerID, Uint32 interval){
 	Game * game = static_cast<Game *>(userdata);
 	game->updatetitle = true;
 	game->fps = game->frames;
