+++ clients/silencer/src/game/session/game_session.h
@@ -18,6 +18,9 @@ void UnloadGame();
 void JoinGame(LobbyGame & lobbygame, char * password);
 void SpectateGame(LobbyGame & lobbygame, char * password);
 void LeaveJoinedGame();
+bool CheckForQuit();
+bool CheckForEndOfGame();
+bool CheckForConnectionLost();
 void ShowDeployMessage();
 void GiveDefaultItems(Player & player);
