+++ clients/silencer/src/game/ui/game_ui_pipeline.h
@@ -19,13 +19,13 @@ class GameUiPipeline
 public:
 explicit GameUiPipeline(Game & game);
 
-void PrepareFrame(Surface & surface);
-void BeginFrame();
-Clay_RenderCommandArray EndFrame();
-void BuildVisible(Surface & surface, float frametime);
+void PrepareClientUiFrame(Surface & surface);
+void BeginPreparedClientUiFrame();
+Clay_RenderCommandArray EndClientUiFrame();
+void BuildVisibleClientUi(Surface & surface, float frametime);
 void DrawInGameWorldInsets(Surface & surface, float frametime);
-void RenderFrame(Surface & surface, float frametime);
-void ResetDeltas();
+void RenderClientUiFrame(Surface & surface, float frametime);
+void ResetUiFrameDeltas();
 bool HasInputTarget();
 void Push(std::unique_ptr<Screen> s);
 void Pop();
